test: 기본 제공 테스트 코드와 커스텀 작성 테스트 분리
- 유닛 테스트는 클래스 별로 분리
- 통합 예외 테스트는 custom 폴더에 작성
